ai_dpr_system/backend/requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
alembic==1.12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pytesseract==0.3.10
Pillow==10.0.1
pdf2image==1.16.3
transformers==4.34.0
torch==2.1.0
scikit-learn==1.3.1
xgboost==2.0.0
pandas==2.1.1
numpy==1.26.0
boto3==1.28.53
httpx==0.25.0
python-dotenv==1.0.0
tenacity==8.2.3
email-validator==2.0.0
Jinja2==3.1.2
pydantic[email]==2.4.2
python-slugify==8.0.1